BC-769: Mandatory field validation
